Add initial pass at proxy server.
See __doc__ on run_server, or do "script/tsapp help".

The goal here is to be able to work on local files that are
currently formatted for use on the remote server without having
to change them, and then once those files are "good" being able to
get them to the server without changes.

In order to be fully effective the proxy will need to proxy PUT
as well as GET, but this just does GET for now.