Add an init command
init creates a new directory with a stub index.html and
an assets directory

This required changes to the config file handling to allow
more information, which cascades changes throughout the server
handling.

Now there can be a .tsapp file in the app dir or at $HOME. The
local one overrides anything not set in HOME.

port, local_host and target_server can be set in .tsapp, along
with auth_token.