Add a push command
The command takes an argument which is the name of the target
space, _public will be appended to make a bag name.

Files that end in html will have the extension removed upon put.
other files will be PUT as is.

In the process, some of the http write handing was moved out to its
own module. The get handling should go there too.

Also in the process things got a bit sloppy, so some cleanup would
be useful.

Still staying within the stdlib.