If a GET response includes an ETag, pass it on.
fix #5

This is not a complete fix, intentionally. A more global fix
would provide a harness for passing along all the relevant headers,
or just _all_ the headers, but I find it a useful research tool
to know what bugs the missing headers cause. In this case, missing
ETags led to the edit progam not being able to properly reveal a
bug that only shows up when edit is installed for real, and not
run through tsapp:

  https://github.com/cdent/edit/commit/4bd915d2c6359d8b1a4e1acbe8c08680160f20a5
  https://github.com/TiddlySpace/tiddlyspace/issues/1014

The code that is in place now provides an opportunity to add
headers as needed, rather than wholesale.

This is one of my twisted ways of doing both personal and community-based
learning which probably enrages someone like @FND.