Switch to using cherrypy as the default server.
Transcontinental latency is too high for the singlethreaded
nature of the wsgiref server: timeouts can happen. We'd like
the default experience to be more positive.

Getting this to work means adding a bit of logging middleware
and adding the cherrpy dependency. That means we now have
an outside of stdlib dependency, but pip works...