If we cannot guess the type of local file, fallback to JSON
This is an experimental feature that may need to be reverted
but it seems like it may be a useful hack.

In the past if the type could not be guessed we'd get an
error from the wsgi server. Now, intsead, fallback to application/json.

This makes it possible to drop JSON data files in the assets dir
for local testing. This can be useful when you want to load a tiddler
from local instead of remote. Usually we wouldn't expect these
files to get pushed. In fact the only way I've used it so far
is in a tsapp that _never_ gets pushed.